Feature/patient

+++ Backend-AuraNeuro.API/Patient/Domain/Command/AssignNeurologistToPatientCommand.cs
@@ -1,3 +1,3 @@
 namespace Backend_AuraNeuro.API.Patient.Domain.Command;
 
-public record AssignNeurologistToPatientCommand(int PatientId, long NeurologistId);
+public record AssignNeurologistToPatientCommand(long PatientId, long NeurologistId);
