+++ Backend-AuraNeuro.API/Patient/Domain/Command/RemoveNeurologistFromPatientCommand.cs
@@ -1,3 +1,3 @@
 namespace Backend_AuraNeuro.API.Patient.Domain.Command;
 
-public record RemoveNeurologistFromPatientCommand(int PatientId, long NeurologistId);
+public record RemoveNeurologistFromPatientCommand(long PatientId, long NeurologistId);
