+++ Backend-AuraNeuro.API/Patient/Domain/Queries/GetPatientByIdQuery.cs
@@ -3,4 +3,4 @@ namespace Backend_AuraNeuro.API.Patient.Domain.Queries;
 /// <summary>
 /// Query to get a patient by their Id.
 /// </summary>
-public record GetPatientByIdQuery(int PatientId);
+public record GetPatientByIdQuery(long PatientId);
